Reorganize the artists onboarding step component
