+++ src/v2/Components/Onboarding/ArtistsStep/ArtistSearchResults.tsx
@@ -1,9 +1,9 @@
-import { PopularArtists_popular_artists } from "v2/__generated__/PopularArtists_popular_artists.graphql"
+import { ArtistSearchResults_viewer } from "v2/__generated__/ArtistSearchResults_viewer.graphql"
 import {
-  PopularArtistsFollowArtistMutation,
-  PopularArtistsFollowArtistMutationResponse,
-} from "v2/__generated__/PopularArtistsFollowArtistMutation.graphql"
-import { PopularArtistsQuery } from "v2/__generated__/PopularArtistsQuery.graphql"
+  ArtistSearchResultsArtistMutation,
+  ArtistSearchResultsArtistMutationResponse,
+} from "v2/__generated__/ArtistSearchResultsArtistMutation.graphql"
+import { ArtistSearchResultsQuery } from "v2/__generated__/ArtistSearchResultsQuery.graphql"
 import { SystemContextProps, withSystemContext } from "v2/Artsy"
 import { SystemQueryRenderer as QueryRenderer } from "v2/Artsy/Relay/SystemQueryRenderer"
 import React from "react"
@@ -15,41 +15,41 @@ import {
 } from "react-relay"
 import track, { TrackingProp } from "react-tracking"
 import { RecordSourceSelectorProxy } from "relay-runtime"
-import { get } from "v2/Utils/get"
-import Events from "../../../../Utils/Events"
-import ReplaceTransition from "../../../Animation/ReplaceTransition"
-import ItemLink, { LinkContainer } from "../../ItemLink"
-import { FollowProps } from "../../Types"
+import Events from "../../../Utils/Events"
+import ReplaceTransition from "../../Animation/ReplaceTransition"
+import ItemLink, { LinkContainer } from "../ItemLink"
+import { FollowProps } from "../Types"
 
-type Artist = PopularArtists_popular_artists[number]
+type Artist = ArtistSearchResults_viewer["searchConnection"]["edges"][number]["node"]
 
-export interface RelayProps {
+export interface ContainerProps extends FollowProps {
+  term: string
   tracking?: TrackingProp
-  relay?: RelayProp
-  popular_artists: PopularArtists_popular_artists
 }
 
-interface Props
-  extends React.HTMLProps<HTMLAnchorElement>,
-    RelayProps,
-    FollowProps {}
+interface Props extends React.HTMLProps<HTMLAnchorElement>, ContainerProps {
+  relay?: RelayProp
+  viewer: ArtistSearchResults_viewer
+}
 
 @track({}, { dispatch: data => Events.postEvent(data) })
-class PopularArtistsContent extends React.Component<Props, null> {
+class ArtistSearchResultsContent extends React.Component<Props, null> {
   private excludedArtistIds: Set<string>
   followCount: number = 0
 
   constructor(props: Props, context: any) {
     super(props, context)
     this.excludedArtistIds = new Set(
-      this.props.popular_artists.filter(Boolean).map(item => item.internalID)
+      this.props.viewer.searchConnection.edges.map(
+        ({ node }) => node.internalID
+      )
     )
   }
 
   onArtistFollowed(
     artist: Artist,
     store: RecordSourceSelectorProxy,
-    data: PopularArtistsFollowArtistMutationResponse
+    data: ArtistSearchResultsArtistMutationResponse
   ): void {
     const suggestedArtistEdge =
       data.followArtist.artist.related.suggestedConnection.edges[0]
@@ -59,21 +59,20 @@ class PopularArtistsContent extends React.Component<Props, null> {
     )
     this.excludedArtistIds.add(artistToSuggest.getValue("internalID") as string)
 
-    const popularArtistsRootField = store.get("client:root")
-    const popularArtists =
-      popularArtistsRootField.getLinkedRecords("popular_artists", {
-        exclude_followed_artists: true,
-      }) || []
+    const popularArtistsRootField = store.get("client:root:viewer")
+    const popularArtists = popularArtistsRootField.getLinkedRecords(
+      "match_artist",
+      { term: this.props.term }
+    )
+    const updatedPopularArtists = popularArtists.map(artistItem =>
+      artistItem.getDataID() === artist.id ? artistToSuggest : artistItem
+    )
 
-    const updatedPopularArtists = popularArtists
-      .filter(Boolean)
-      .map(artistItem =>
-        artistItem.getDataID() === artist.id ? artistToSuggest : artistItem
-      )
-
-    store
-      .get("client:root")
-      .setLinkedRecords(updatedPopularArtists, "popular_artists")
+    popularArtistsRootField.setLinkedRecords(
+      updatedPopularArtists,
+      "match_artist",
+      { term: this.props.term }
+    )
 
     this.followCount += 1
 
@@ -83,17 +82,17 @@ class PopularArtistsContent extends React.Component<Props, null> {
       action: "Followed Artist",
       entity_id: artist.internalID,
       entity_slug: artist.slug,
-      context_module: "onboarding recommended",
+      context_module: "onboarding search",
     })
   }
 
   onFollowedArtist(artist: Artist) {
-    commitMutation<PopularArtistsFollowArtistMutation>(
+    commitMutation<ArtistSearchResultsArtistMutation>(
       this.props.relay.environment,
       {
         // TODO: Inputs to the mutation might have changed case of the keys!
         mutation: graphql`
-          mutation PopularArtistsFollowArtistMutation(
+          mutation ArtistSearchResultsArtistMutation(
             $input: FollowArtistInput!
             $excludedArtistIds: [String]!
           ) {
@@ -103,7 +102,6 @@ class PopularArtistsContent extends React.Component<Props, null> {
                 excludeFollowedArtists: true
                 excludeArtistIDs: $excludedArtistIds
               ) {
-                slug
                 internalID
                 id
                 name
@@ -123,7 +121,6 @@ class PopularArtistsContent extends React.Component<Props, null> {
                   ) {
                     edges {
                       node {
-                        slug
                         internalID
                         id
                         name
@@ -153,12 +150,10 @@ class PopularArtistsContent extends React.Component<Props, null> {
   }
 
   render() {
-    const artistItems = this.props.popular_artists
-      .filter(Boolean)
-      .map((artist, index) => {
-        const imageUrl = get(artist, a => a.image.cropped.url)
+    const artistItems = this.props.viewer.searchConnection.edges.map(
+      ({ node: artist }, index) => {
         return (
-          <LinkContainer key={`popular-artists-${index}`}>
+          <LinkContainer key={`artist-search-results-${index}`}>
             <ReplaceTransition
               transitionEnterTimeout={1000}
               transitionLeaveTimeout={400}
@@ -167,32 +162,42 @@ class PopularArtistsContent extends React.Component<Props, null> {
                 href="#"
                 item={artist}
                 key={artist.id}
-                id={artist.id}
-                name={artist.name}
-                image_url={imageUrl}
+                id={artist.internalID}
+                name={artist.displayLabel}
+                image_url={artist.imageUrl}
                 onClick={() => this.onFollowedArtist(artist)}
               />
             </ReplaceTransition>
           </LinkContainer>
         )
-      })
+      }
+    )
 
     return <div>{artistItems}</div>
   }
 }
 
-const PopularArtistContentContainer = createFragmentContainer(
-  PopularArtistsContent,
+const ArtistSearchResultsContentContainer = createFragmentContainer(
+  ArtistSearchResultsContent,
   {
-    popular_artists: graphql`
-      fragment PopularArtists_popular_artists on Artist @relay(plural: true) {
-        slug
-        internalID
-        id
-        name
-        image {
-          cropped(width: 100, height: 100) {
-            url
+    viewer: graphql`
+      fragment ArtistSearchResults_viewer on Viewer {
+        searchConnection(
+          query: $term
+          mode: AUTOSUGGEST
+          entities: [ARTIST]
+          first: 10
+        ) {
+          edges {
+            node {
+              ... on SearchableItem {
+                id
+                slug
+                internalID
+                displayLabel
+                imageUrl
+              }
+            }
           }
         }
       }
@@ -200,28 +205,26 @@ const PopularArtistContentContainer = createFragmentContainer(
   }
 )
 
-const PopularArtistsComponent: React.SFC<SystemContextProps & FollowProps> = ({
-  relayEnvironment,
-  updateFollowCount,
-}) => {
+const ArtistSearchResultsComponent: React.SFC<
+  ContainerProps & SystemContextProps
+> = ({ term, relayEnvironment, updateFollowCount }) => {
   return (
-    <QueryRenderer<PopularArtistsQuery>
+    <QueryRenderer<ArtistSearchResultsQuery>
       environment={relayEnvironment}
       query={graphql`
-        query PopularArtistsQuery {
-          highlights {
-            popular_artists: popularArtists(excludeFollowedArtists: true) {
-              ...PopularArtists_popular_artists
-            }
+        query ArtistSearchResultsQuery($term: String!) {
+          viewer {
+            ...ArtistSearchResults_viewer
           }
         }
       `}
-      variables={{}}
+      variables={{ term }}
       render={({ error, props }) => {
         if (props) {
           return (
-            <PopularArtistContentContainer
-              popular_artists={props.highlights.popular_artists}
+            <ArtistSearchResultsContentContainer
+              viewer={props.viewer}
+              term={term}
               updateFollowCount={updateFollowCount}
             />
           )
@@ -233,4 +236,6 @@ const PopularArtistsComponent: React.SFC<SystemContextProps & FollowProps> = ({
   )
 }
 
-export const PopularArtists = withSystemContext(PopularArtistsComponent)
+export const ArtistSearchResults = withSystemContext(
+  ArtistSearchResultsComponent
+)
