+++ src/v2/Components/Onboarding/ArtistsStep/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ArtistsStep"
